Fix observer & fix some index.html / webpack issues
