+++ Behavioral/Memento/index.html
@@ -5,7 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width,initial-scale=1" />
     <title>Memento Pattern: Example</title>
-    <script src="./dist/index.js" defer="defer"></script>
     <link rel="stylesheet" href="style.css" />
     <script defer="defer" src="dist/index.js"></script>
   </head>
