+++ Behavioral/Observer/index.html
@@ -1,63 +1,1 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Observer Pattern: Example</title>
-    <script src="./dist/observerPattern.js" defer></script>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body style="overflow-x: hidden">
-    <div class="server-application">
-      <h3 style="text-align: center">Server</h3>
-      <div class="window-holder">
-        <div
-          class="window"
-          style="display: inline-block; overflow-y: auto; text-align: center"
-        >
-          <nav></nav>
-          <h2>SERVER | DATABASE</h2>
-          <ul class="time-window"></ul>
-        </div>
-        <div class="window">
-          <nav></nav>
-          <h2>LOG</h2>
-          <ul class="log-updates"></ul>
-        </div>
-      </div>
-    </div>
-    <div class="user-application">
-      <h3 style="text-align: center">Users</h3>
-      <div class="window-holder">
-        <div class="window user1">
-          <nav></nav>
-          <h2>USER 1</h2>
-          <div class="user-info">
-            <h5>System LOG</h5>
-            <ul class="user-log-news user1"></ul>
-          </div>
-          <button type="button" class="user1-button">Log Now</button>
-        </div>
-        <div class="window user2">
-          <nav></nav>
-          <h2>USER 2</h2>
-          <div class="user-info">
-            <h5>System LOG</h5>
-            <ul class="user-log-news user2"></ul>
-          </div>
-          <button type="button" class="user2-button">Log Now</button>
-        </div>
-        <div class="window user3">
-          <nav></nav>
-          <h2>USER 3</h2>
-          <div class="user-info">
-            <h5>System LOG</h5>
-            <ul class="user-log-news user3"></ul>
-          </div>
-          <button type="button" class="user3-button">Log Now</button>
-        </div>
-      </div>
-    </div>
-  </body>
-</html>
+<!doctype html><html lang="en"><head><meta charset="UTF-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width,initial-scale=1"/><title>Observer Pattern: Example</title><link rel="stylesheet" href="style.css"/><script defer="defer" src="dist/index.js"></script></head><body style="overflow-x: hidden"><div class="server-application"><h3 style="text-align: center">Server</h3><div class="window-holder"><div class="window" style="display: inline-block; overflow-y: auto; text-align: center"><nav></nav><h2>SERVER | DATABASE</h2><ul class="time-window"></ul></div><div class="window"><nav></nav><h2>LOG</h2><ul class="log-updates"></ul></div></div></div><div class="user-application"><h3 style="text-align: center">Users</h3><div class="window-holder"><div class="window user1"><nav></nav><h2>USER 1</h2><div class="user-info"><h5>System LOG</h5><ul class="user-log-news user1"></ul></div><button type="button" class="user1-button">Log Now</button></div><div class="window user2"><nav></nav><h2>USER 2</h2><div class="user-info"><h5>System LOG</h5><ul class="user-log-news user2"></ul></div><button type="button" class="user2-button">Log Now</button></div><div class="window user3"><nav></nav><h2>USER 3</h2><div class="user-info"><h5>System LOG</h5><ul class="user-log-news user3"></ul></div><button type="button" class="user3-button">Log Now</button></div></div></div></body></html>
