+++ Behavioral/Observer/src/models/Observer.ts
@@ -1,0 +1,3 @@
+export interface Observer {
+  update(): void;
+}
