+++ Behavioral/Observer/src/models/Subject.ts
@@ -1,0 +1,7 @@
+import { Observer } from "./Observer";
+
+export interface Subject {
+  add(observer: Observer): void;
+  remove(observer: Observer): void;
+  notify(): void;
+}
